+++ addons/tents/XEH_PREP.hpp
@@ -1,0 +1,8 @@
+PREP(addActions);
+PREP(cancel);
+PREP(canPackUp);
+PREP(construct);
+PREP(packUp);
+PREP(place);
+PREP(progressPackUp);
+PREP(startBuild);