+++ addons/tents_bwa3/script_component.hpp
@@ -1,0 +1,13 @@
+#define COMPONENT tents_bwa3
+#define COMPONENT_BEAUTIFIED Tents (BWA3)
+#include "\z\avo\addons\main\script_mod.hpp"
+
+#ifdef DEBUG_ENABLED_TENTS_BWA3
+    #define DEBUG_MODE_FULL
+#endif
+
+#ifdef DEBUG_SETTINGS_TENTS_BWA3
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_TENTS_BWA3
+#endif
+
+#include "\z\avo\addons\main\script_macros.hpp"